test return info if timeout

+++ src/OfficeConverter/OfficeConverter.php
@@ -50,18 +50,21 @@ class OfficeConverter
 
         $outdir = $this->tempPath;
         $shell = $this->exec($this->makeCommand($outdir, $outputExtension, $timeout));
-        // echo '<pre>$shell : ', print_r($shell, true) ,'</pre>';
+        echo '<pre>$shell : ', print_r($shell, true) ,'</pre>';
+        $shell['return'] = 124;
         if (0 != $shell['return']) {
-            // echo '<pre>info : ', print_r([
-            //     $filename,
-            //     file_exists($outdir.'/'.$filename) ? "Exists" : " Does not exist",
-            //     '.~lock.'.$filename.'#',
-            //     file_exists($outdir. '/.~lock.'.$filename.'#') ? "Exists" : " Does not exist",
-            //     $outdir,
-            //     $outputExtension
-            // ], true) ,'</pre>';
+            
+            if (124 != $shell['return']) {
+                echo '<pre>info : ', print_r([
+                    $filename,
+                    $outdir,
+                    $outputExtension
+                ], true) ,'</pre>';
 
-            if (file_exists($outdir. '/.~lock.'.$filename.'#')) unlink($outdir. '/.~lock.'.$filename.'#');
+                if (file_exists($outdir. '/.~lock.'.$filename.'#')) unlink($outdir. '/.~lock.'.$filename.'#');
+
+                throw new OfficeConverterException('Convertion Timed Out! Contact Server Admin.');
+            }
 
             throw new OfficeConverterException('Convertion Failure! Contact Server Admin.');
         }
